fix error position on tvars in logical operators
Summary:we were missing repositioning when refining logical operators, which meant that tvars caused "errors at a distance" (the value's definition location, rather than where it's used in the logical operator).

the most consistent way to do this would be to reposition to the location of the entire logical expresion; indeed, the entire expression is what doesn't match, but it can be hard to tell which branch is the problem. so as a compromise, I'm tracking the left and right locations and using the correct one. this only works one level deep (`x && y && z` makes us point at `x && y` where we used to point at `x`, for example... see the updated tests). in the future, I'd like to undo this so it does point at the entire expression and then says "i think this expression is a string because <the left side> could be a string; I think <the left side> could be a string because <its left side> could be a string."

Reviewed By: gabelevi

Differential Revision: D3195968

fb-gh-sync-id: e4930381e00e9b40346d959e0d88440b736c2872
fbshipit-source-id: e4930381e00e9b40346d959e0d88440b736c2872