Make flow automatically load libs from `flow-typed`
Summary:We've been organizing an effort over at https://github.com/flowtype/flow-typed to build a repository of 3rd party library definitions.

As part of this effort, jeffmo and I thought it'd be a good idea to have flow automatically load any definitions from the `flow-typed` folder if it exists.
Closes https://github.com/facebook/flow/pull/1640

Reviewed By: gabelevi

Differential Revision: D3164756

Pulled By: jeffmo

fb-gh-sync-id: e7e2bc4506dfffb5290cac2aff0383057bc79d47
fbshipit-source-id: e7e2bc4506dfffb5290cac2aff0383057bc79d47