store repl content in URL fragment
Summary:adds permalinks to /try by storing the contents of the file in the URL fragment as you type.

this is cool because:

- the file content is entirely client-side. it's never transmitted to our servers.
- no DB required
- updates as you type. no permalink/share button needed

this is not cool because:

- URL length limits. IE is limited to 2083 characters, which you can hit super quickly. so, this doesn't work great in IE, but I don't think IE is super widely used by our audience.
- metadata: if we support multiple files, or config like choosing flowconfig options or which libs are available, it might get harder to serialize like this.

if we outgrow this approach, I think we can use the GitHub API to create gists, or save things to dynamoDB. but that's a lot more complicated and doesn't update as you type!

Reviewed By: gabelevi, samwgoldman

Differential Revision: D3210845

fb-gh-sync-id: 2d45cdd1029b7f7b557e705d4ef9c0e10c1016de
fbshipit-source-id: 2d45cdd1029b7f7b557e705d4ef9c0e10c1016de