Fix --strip-root pretty printing
Summary:Pretty printing errors with `--strip-root` would only work when run
from the root directory. This is because we treated those paths as relative
paths from the current working directory, rather than from the root, so we
couldn't read from the referred to files.

Easy fix - just need to pass the root into the function that reads from the
source files.

This is why the new test runner was outputting weird output - context was
missing from the json output.

Reviewed By: avikchaudhuri

Differential Revision: D3213062

fb-gh-sync-id: cbf1e77b47b30816eab6db71d4fccb6b668244b8
fbshipit-source-id: cbf1e77b47b30816eab6db71d4fccb6b668244b8