./tool record (3 of 5)
Summary:So a test suite is made up of tests. A test is made up of steps. Each
step is made up actions and assertions.

When an assertion fails, it saves a suggestion. `tool record myTest` runs
`myTest` and then applies the suggestions. So if you change a lib file and 100
tests fail, you can just run `./tool record` and all the failing tests will be
re-recorded.

`./tool help record`:

  usage: tool record [OPTION]... [SUITE]...

  SUITE
      Given the test `path/to/testdir/mytest/test.js`, the following values of
      SUITE will record that test:
          * mytest
          * path/to/testdir/mytest
          * path/to/testdir/mytest/test.js

      If no suites are specified, every test in the test directory will be recorded.

  OPTIONS:
      --bin path/to/flow
          Path to the flow binary
      --check COMMAND, -c COMMAND
          The flow command to check flow errors ["check", "status"] (default: "check")
      --help, -h
          Shows this usage message
      --parallelism N, -p N
          Number of tests to run in parallel (default: "16")

Reviewed By: jeffmo

Differential Revision: D3085934

fb-gh-sync-id: 5cfb146cd0b9c6636218edc6a1a3641187a168e9
fbshipit-source-id: 5cfb146cd0b9c6636218edc6a1a3641187a168e9