./tool convert (4 of 5)
Summary:This is a pretty hacky command that tries to convert old tests into
new tests. It only works for a certain type of test, but that's most of our
tests.

Most of our old tests have a bunch of files in a directory, run `flow check`,
and compare the output to a `foo.exp` file. This converts cleanly to a new test
which does exactly that.

The convert command has 3 steps:

1. Roughly create a new test `legacy/foo` from the old test `foo`
2. Run `./tool record newtests/legacy/foo`
3. Do some rough sanity check

The first step basically just copies the files from `tests/foo` to
`newtests/legacy/foo`. It renames a few files, creates a `test.js` from a
template, and puts placeholders for the assertions.

The second step just runs `./tool record`

The third step is a really rough sanity check. It looks at how many errors the
old test had and how many errors the new test has. If these are equal then it's
done.

`./tool help convert`
  usage: /data/users/glevi/fbsource/fbcode/flow/tool convert [OPTION]... LEGACY_TEST_DIR...

  LEGACY_TEST_DIR
      This is the path to the legacy test directory. Given
      `path/to/legacy/tests/mytest`, convert will create
      `path/to/new/tests/legacy/mytest`. At least one
      legacy test must be provided.

  OPTIONS:
      --bin path/to/flow
          Path to the flow binary
      --check COMMAND, -c COMMAND
          The flow command to check flow errors ["check", "status"] (default: "check")
      --help, -h
          Shows this usage message
      --parallelism N, -p N
          Number of tests to run in parallel (default: "16")
      --sanity
          Don't convert, just run sanity check

Reviewed By: jeffmo

Differential Revision: D3086095

fb-gh-sync-id: fbe7927f34ca5b11eceb47f8c6d4582813cb3156
fbshipit-source-id: fbe7927f34ca5b11eceb47f8c6d4582813cb3156