avoid naive unification on PolyT ~> PolyT
Summary:
Previously, we did naive unification (bidirectional subtype checking)
on `PolyT ~> PolyT`. This could cause an explosion of test cases,
especially in nested situations such as polymorphic classes hosting
polymorphic methods. And since checking such a definiion against
itself is not uncommon, it wasn't that hard to provoke the pathology
(see new test for an example that effectively hangs the typecheck).

This diff captures a subset of such cases, where the lower and upper
`PolyT`s have the same arity. In those cases it unifies the upper bounds
of each pair of type params, then instantiates each poly type with the
upper bounds as arguments, and unifies the pair. The idea is that these
unifications should be sufficient to check structural congruence.
(This leaves other easy checks unexploited, such as arity mismatches,
in the interests of keeping the diff relatively tiny. :)

Fixes github issue https://github.com/facebook/flow/issues/1029

Reviewed By: avikchaudhuri

Differential Revision: D3218650

fb-gh-sync-id: ecd1bfe831d82f225a282f362c0d3b983543c36a
fbshipit-source-id: ecd1bfe831d82f225a282f362c0d3b983543c36a