compress fragments in /try/ URLs
Summary:
uses [lz-string](http://pieroxy.net/blog/pages/lz-string) to compress the code before base64 encoding it.

also adds a version number as the first character of the fragment, so that we can change the encoding in the future and maintain backwards compatibility. this gives us space for 63 versions before we need to add another digit (when frag[0] === 'Z', then version = 64 + frag[1], for example).

Reviewed By: samwgoldman

Differential Revision: D3222830

fb-gh-sync-id: 7ae2dcd05cb53e2fa77cd727746c65da2fac60ff
fbshipit-source-id: 7ae2dcd05cb53e2fa77cd727746c65da2fac60ff