Remove `is_void` function return type munging
Summary:
This original intent of this code was to remove the annotation requirement for
functions with explicitly void return types.

Now that return types are not required anywhere, this code is superfluous. I
have removed `is_void` and its callers. I have also removed `Flow.check_type`
which was only called through `is_void`.

It turns out that this code was responsible for an apparently long-standing bug
where the inferred return type for functions which *may* return `void` was
exactly `void`, rather than a union of all possible return types.

Reviewed By: avikchaudhuri, bhosmer, mroch

Differential Revision: D3222927

fb-gh-sync-id: f38f65e804284c0693bbc0a7a485ec2d487d74aa
fbshipit-source-id: f38f65e804284c0693bbc0a7a485ec2d487d74aa