fix incremental mode when JSON changes
Summary:
we were mis-classifying `.json` paths as `SourceFile` on updates, so none of the maps/sets containing `Loc.filename`s matched `JsonFile filename.json`.

this caused a crash when the server was started with an existing .json file, and then that json file was deleted. recheck didn't realize the file was deleted properly, causing it to go through the recheck flow which fataled.

Reviewed By: samwgoldman

Differential Revision: D3225582

fb-gh-sync-id: 179e808bc01aa708464062ee6b4ef0ade2f22c3b
fbshipit-source-id: 179e808bc01aa708464062ee6b4ef0ade2f22c3b