special forms of UseT
Summary:
"This type is incompatible with" errors have the potential to be a lot more descriptive, if we can keep track of what caused us to flow one def type into another def type.

This diff is a lot of boilerplate nonsense to add an "op" property to `UseT`, and a couple of examples of those ops being set when we deconstruct some other type into the final `Type.t -> UseT Type.t`.

Reviewed By: avikchaudhuri

Differential Revision: D3208938

fb-gh-sync-id: 4fc3fafb4a01d4bdf8d490a71ab99e2975ceb76f
fbshipit-source-id: 4fc3fafb4a01d4bdf8d490a71ab99e2975ceb76f