Use Class_sig (a.k.a. Inst_sig) for instances, too
Summary:
Classes and interfaces have a lot in common. Correspondingly, their AST
traversals shared a lot of code. This is explicitly called out for `mk_instance`
and `mk_class`, but the `interface` function within `statement` also shares a
lot in common with `mk_class_signature`.

Generalizing Class_sig a tiny bit to model overloaded methods, we can reuse the
same code for both. I suspect there is even more opportunity for re-use, but
this seemed like a good chunk of complexity to put up for review.

Reviewed By: gabelevi

Differential Revision: D3240179

fb-gh-sync-id: 6961eada0d6ce88c3075111c341313cb904f5e5a
fbshipit-source-id: 6961eada0d6ce88c3075111c341313cb904f5e5a