[hack] Retrieve method extents position in ServerSymbolDefinition
Summary: See the changes to ServerSymbolDefinition.go - now it returns two positions, the position of the identifier, and the extents of entire thing. It's currently ignored value. Because ServerSymbolDefinition requires parsing heap to compute it, we have to populate/restore it too in ServerIdeUtils, in addition to all the other heaps there.

Reviewed By: int3

Differential Revision: D3234660

fb-gh-sync-id: 3b72a8a0f4e6f3ef37caeb710b50090a4edf6346
fbshipit-source-id: 3b72a8a0f4e6f3ef37caeb710b50090a4edf6346