Fix type param substitution for function signatures
Summary:
Re-using Class_sig for interfaces and declared classes exposed a bug, which was
in fact a long-standing bug that always existed for classes, but was never
properly diagnosed.

We were mishandling type parameters for methods with the same name as a type
parameter from its class. The expected behavior is that the method's type
parameter "shadow" the class's within the body of the method.

To make things more confusing, we did this correctly in one place, but not
another. In mk_body, we used the correct type_params_map, but the generated
InstanceT was wrong. This caused the symptoms to only appear when there was also
a SuperT constraint, or a SetElemT/GetElemT.

This commit fixes the underlying issue, so now classes and interfaces (and
declared classes) work correctly, with respect to this behavior at least.

Reviewed By: gabelevi

Differential Revision: D3252342

fb-gh-sync-id: a7902d654ed5d4c5fc2374cc46ec61c4fcebab9f
fbshipit-source-id: a7902d654ed5d4c5fc2374cc46ec61c4fcebab9f