Down with `declare var exports`, all hail `declare module.exports`!
Summary:
tldr: `declare var exports` is hacky and deserves it's own syntax. This diff creates a `declare module.exports` syntax to express the intent better.

== Why?

Today, the only way to declare the “clobbered” type of module.exports for a CommonJS module today is by using a trick where one declares a value export named exports. This is undiscoverable and unintuitive (both from the perspective of the person writing the `declare module` and the person reading it).

It also gives a muddled illusion of providing a default export for a modeled ES module, but in fact it does not (it defines a named export called default — by way of CommonJS↔ES interop).

== Fate of `declare var exports`

This diff leaves the semantics of `declare var exports` in tact alongside `declare var module.exports` to support a grace period for transition. If both are present in a `declare module`, `declare module.exports` takes precedence (and `declare var exports` reverts to it's future behavior of doing what all other `declare var`s do).

After the transition period (hopefully just 1 or 2 releases), I intend to remove support for `declare var exports` and step into the shiny new world of cleanliness.

This diff is the first in a series of diffs to clean up `declare module`.

Reviewed By: gabelevi

Differential Revision: D3248255

fb-gh-sync-id: fd8b97f295f018b605a108fdaab63cf79e7a8767
fbshipit-source-id: fd8b97f295f018b605a108fdaab63cf79e7a8767