Fix Makefile rebuilding of object files
Summary:
I was having problems with `make` not picking up changes to `.c` files. I think the problem is that `build-flow` depends on `build-flow-native-deps` which does not actually depend on the `.c` files, so that step was getting skipped.

As I started tweaking stuff, I eventually just replaced the target that runs `ocamlbuild` on all `.o` files with a rule that directly runs `ocamlopt` on each `.c` file, letting `make` instead of `ocamlbuild` decide when it's dirty. this is faster.
Closes https://github.com/facebook/flow/pull/1753

Reviewed By: gabelevi

Differential Revision: D3270015

Pulled By: mroch

fb-gh-sync-id: 3515c883a1f408b9807a7446b237b8a01415ea74
fbshipit-source-id: 3515c883a1f408b9807a7446b237b8a01415ea74