[hack] Make decl naming output consistent
Summary:
I would like to kill the Nast heap. Decl-naming doesn't do a lot of
CPU-intensive work, so there is little benefit in storing the output of that
step in shared memory. Instead we can recompute the Nast from the AST each time
we need it.

However, running decl-naming multiple times over the same input doesn't produce
the same results. In particular, the param_id changes, since the function /
method parameters get mapped to integer IDs using a global counter. There are
a number of places in the code that assume that the function parameters get
mapped to the same IDs each time, so we cannot break this invariant.

This diff solves the problem by hashing the string representation of the params
to obtain the integer IDs, thus guaranteeing reproducibility. This also means
that there is no longer a need to store the integer ID of the params in the
Nast, since they can be recomputed cheaply.

Certainly, hashing is not the only way to get reproducible IDs. We could also
number the parameters starting from zero. However, we need to ensure that it
doesn't overlap with the IDs we assign to local variables in the function body.
Since we name the bodies in a separate phase from the parameters, this is
somewhat inconvenient. Hashing allows us to avoid carrying around a bunch
of extra state.

This refactor also means that there is no longer a need to use a global
shared memory counter. Local_id can now live under `utils/` rather than `heap/`.

This change does break `--dump-symbol-info` a little -- ident numbers are now
only unique on a per-method basis instead of a per-file basis. I've added a test
to keep track of this regression, but it's not the highest priority to fix. I think I
also fixed a latent bug in this flag -- previously, the usages of a method param
would have a different ident than the param itself; now they are identical.

Reviewed By: dabek

Differential Revision: D3241207

fb-gh-sync-id: cd911ae48d633f18caa71223d1017a6753b73ab2
fbshipit-source-id: cd911ae48d633f18caa71223d1017a6753b73ab2