Separate constructor in Class_sig
Summary:
Just a small cleanup here, but the payoff doesn't come until the next diff. It's
fine to treat constructors as methods, except that we want to some extra stuff
about them.

Currently we ask questions like "should we disable `this` and `super`?" for
derived constructors. We ask this question for all methods, but it only makes
sense for constructors.

In the future we will want to ask questions like "were all the declared fields
initialized?" and while this (and subsequent) diffs don't answer that question,
widening the model will hopefully pave the way.

Reviewed By: gabelevi

Differential Revision: D3262047

fbshipit-source-id: 30ba88993c98ecfed47dc922e9577b3605a75682