Add .mli files with documentation for public API
Summary:
Now that this refactoring work is done, I wanted to make sure the public API is
clear.

In the process of adding the .mli, I realized I didn't fully encapsulate
Class_sig.t, but I was able to trivially fix that by pushing `mutually` and
`ignore` into `check_super`.

Reviewed By: gabelevi

Differential Revision: D3273507

fbshipit-source-id: 12f2e4d590227069ef05bb56712c057ba15bcefa