Parser Tests
Summary:
This fixes most parser Tests.
Left are the Decorator related hardcoded tests.
This is because ast-types has a special node for Decorators. Flow does not.
I will make another PR that parses the Decorators correctly as Decorator nodes.

But thats a seperate issue. This was mainly to update the tests and ast-types.
Closes https://github.com/facebook/flow/pull/1717

Reviewed By: samwgoldman

Differential Revision: D3265034

Pulled By: gabelevi

fbshipit-source-id: b6d24e486b1ea0049bbc6d4ed8c8a5e8f42d59e3