[parser] remove Expect.eof
Summary:
This appears to be identical to `Expect.token env T_EOF`. the only difference is that `lex_env env` is now called before `lookahead env`, but `lookahead` does not appear to mutate `env.lex_env`.

Gets rid of a caller of `Eat.advance` which `Parser_env` calls an abomination.

Reviewed By: gabelevi

Differential Revision: D3285125

fbshipit-source-id: f420436fb3b7be80e351d2acb689fc0533cd4c10