fix parsing of `/x/* 5 */y/`
Summary:
we correctly saw that `/x/` is a regex, `*` is a multiplication and `5` is a number. then we think `*/` is closing a comment that isn't open, and error.

normally, we would only look for `*/` while parsing a comment (in fact, this is what the `comment` rule does). however, comment-syntax mode (`/*:: ... */`) requires parsing arbitrary tokens until the closing `*/`, so we had a regex for `*/` that matched even when it should really be 2 tokens.

Reviewed By: gabelevi

Differential Revision: D3286928

fbshipit-source-id: c7cf965a74f20907619db10af1b641ba46cbaffb