[hack] Docblocks in outline view
Summary:
With information from current parser this is mostly a "best effort" heuristic, but seems to work well enough to make it worth exposing.

What is slightly annoying is that ideally we would want to expose this information in more (than outline view) time-sensitive contexts, like autocomplete suggestions on identify function hovercards. But getting the docblock requires reparsing (or at least re-lexing) the file, which in turn requires hitting the disk - so far we've been staying away from accessing the disk at any moment outside of parsing stage so I'm a little vary of crossing this bridge right now.

Reviewed By: int3

Differential Revision: D3290860

fbshipit-source-id: 5c556979d35bfdd099e5cfa2fff217ebf6690ddd