[parser] refactor lookahead, eliminating vomit!
Summary:
= Before this diff =

When the lexer is initialized, it lexes the first token and stores it in `env.single_lookahead`. Calling `lookahead env` just returns this token. When the lexer is advanced by calling `advance`, you have to pass in the current token (which is almost always the same as `single_lookahead`); then `lex` is called again to put the next token in `single_lookahead`.

If we peek at the looked-ahead token and decide to change the lexer mode, we have to back up ("vomit") and re-lex in the new mode. Backing up is pretty sketchy. It only works because we have the entire file in the `lexbuf` buffer via `Lexing.from_string`. Normally, `lexbuf.lex_buffer` contains a sliding window, and advancing to the next token (during lookahead) can evict the previous token from the buffer, causing us to try to back up past the beginning of the buffer. This prevents us from trying to use `Lexing.from_channel` to stream the file in (e.g. to only read the head of the file to extract the docblock).

= After this diff =

This diff eliminates `single_lookahead` and instead uses the cloned `lexbuf` that `Lookahead.create` makes. When we want to peek, we do so in the clone so if we need to change modes, we can do so without vomiting by creating a new `Lookahead`. If we accept a looked-ahead token, we commit those changes to the primary `lexbuf` and don't need to re-lex it to move the primary `lexbuf` forward.

Because the lookahead is happening in a clone, regex and template literal parsing are a little more natural. For example, we lookahead and see a `T_DIV` and know a regex is coming up, so we can call `lex_regex` which expects a `/` and generates a `T_REGEXP` without generating a `T_DIV`. we could've achieved the same thing by vomiting and changing modes (if we had a mode for regexes, which we don't), but this seems cleaner.

Reviewed By: gabelevi

Differential Revision: D3297481

fbshipit-source-id: 9758d33b4f52947446bb0f0636050384f2337eb9