Support stateless functional components via CustomFunT
Summary:
This diff uses an internal function type because the "proper" fix, using a union
type for the first parameter of createElement, runs afoul of existing union
bugs.

The internal type waits for the first argument to become concrete, then
dispatches somewhat naively based on the kind of type --- classes are
instantiated, functions are called, strings are looked up in $JSXIntrinsics.

Reviewed By: avikchaudhuri

Differential Revision: D3285681

fbshipit-source-id: 1d2258ef0e561d6e51bfd7f1553008088fe39d7d