fix crash on duplicate methods in class
Summary:
Overwriting duplicate method declarations in a class was not tested. In one
place, we were adding an entry to a map that automatically shadowed any existing
entry; but in another place we did the same to a list, and things went out of
whack, causing a runtime exception later.

Fixed, following the semantics of shadowing for methods and getters/setters.

We can decide later whether to error on duplicates or not. It's probably clearly
unintentional in ES6 classes, although in arbitrary objects containing methods,
shadowing on spreads (to implement mixins) could be intentional.

Reviewed By: samwgoldman

Differential Revision: D3321882

fbshipit-source-id: ad7a5b9c0a3fde11b177e6eae4ebc5aa2b8a62bf