Make ReposUpperT a use_t
Summary:
As a def type, ReposUpperT can run into an OpenT and get into the constraint
graph. If that OpenT gets a lower bound, the ReposUpperT becomes an upper bound,
which shouldn't happen.

Instead of adding delicate code to ensure that a ReposUpperT doesn't run into an
OpenT, I changed it to a use type, which can't. This relies on the convenient
fact that we never actually reposition use types, only upper def types.

The added test in annot2 is the same as the test in annot/any except that the
type T comes from an untyped module instead of an explicit `any`. I needed to
create a new test dir in order to configure `any: false` in the `.testconfig`.

Reviewed By: bhosmer

Differential Revision: D3332872

fbshipit-source-id: 4222175cd0da02a117936e42485f418081d24cf9