Replace AnyWith*BoundT in ElemT with Read/Write variant
Summary:
While working on prop variance, which adds a Read/Write variant for use in
LookupT, I realized that ElemT used AnyWithUpperBoundT and AnyWithLowerBoundT
where Read/Write are more appropriate.

ElemT unwraps the AnyWith*BoundT (with an assert false) in all except one case
which handles dictionaries. I have changed this case in the diff to emit a
single directional constraint instead of two constraints where one is a noop.

Reviewed By: bhosmer

Differential Revision: D3326915

fbshipit-source-id: 6976fb42b8678b904e6a4d3cda01c4a40983f06f