Add some missing regression tests around lookup/shadow prop behavior
Summary:
Lookup operations on object types add shadow properties along the prototype
chain to ensure that any reads and writes on the prototype are compatible, even
if they happen later.

This behavior is pretty subtle. This commit only adds tests to protect against
regressions of the existing behavior. Adding these tests in anticipation of
upcoming changes to LookupT to support prop variance.

Reviewed By: bhosmer

Differential Revision: D3333764

fbshipit-source-id: e27fc0a63c4f414007ce7460fac9493d7d51f809