Allow deferred unification with `any`
Summary:
AnnotT creates a UnifyT constraint between the source and sink, but if the
source is `any`, the special case in `ground_subtypes` prevented the unification
from ever happening.

This led to false negatives, in the form of the added test. The number assigned
to `this.p` in the constructor should never run into the `string` constraint
added in the subclass, but does if we never evaluate the deferred unification
constraint.

(Note: while it's definitely uncommon to specify an any type alias as I have
done in the test, this same behavior happens when importing a type from an
untyped module which is fairly common.)

Reviewed By: avikchaudhuri

Differential Revision: D3332235

fbshipit-source-id: ab8fca33616dae20c0270b0251257ea85d2eba2f