[parser] simplify template_tail lexer
Summary: shuffles some code around to simplify `Lexer_flow.template_tail`. it used to be the only function in the entire lexer that returned an AST node, now that happens in the parser code. This made it more reusable in the lexer, to the point that `template_tail`'s implementation on line 1602 is as simple as the others around it.

Reviewed By: gabelevi

Differential Revision: D3337874

fbshipit-source-id: 269c9124ec3f25999bb81e6addc4a0e78d3b1e5c