[parser] make lex_env immutable
Summary:
most of the lexer functions already returned `lex_env`. this makes the rest do so too, and then changed the functions that add errors and comments to `lex_env.lex_state` to instead return a new `lex_env` instead of mutating it.

doing this to try to help me keep everything straight with lookaheads. storing envs (say, in a `Lookahead` env) that can then be mutated by something else is really confusing me.

Reviewed By: gabelevi

Differential Revision: D3338829

fbshipit-source-id: 3e6b87406ce3cb62e6961a1f63cd6ca531e0d59b