Make workers spawn rather than fork (5 of 5)
Summary:
gabelevi (aka gabelevi) has split
https://github.com/facebook/flow/pull/1184 up into multiple diffs
for to make testing and reviewing easier.

This is the last piece of the pull request and is one of the scariest bits. It
makes the workers use `Daemon.spawn` instead of fork. Now suddenly the workers
don't have access to the global state they had before!

To get all the tests passing I had to make sure to save and restore logging
state.

Reviewed By: int3

Differential Revision: D2796468

Pulled By: gabelevi

fbshipit-source-id: 73eb9dac0431360148dc520601c53702f5598dfb