fix recheck merge crash
Summary:
A bad interaction between reverse dependency calculation and forward dependency
calculation caused merge to crash. In particular, reverse dependency calculation
was failing to include some files in a cycle that could be found by forward dependency
calculation from other files in the cycle.

The specific bug arises because a file's exported module can be referenced in
two ways: by global name or by path. In particular, in a mixed Haste/Node world,
the same file can be referenced both ways in a cycle, and failing to account for
both possibilities might cause us to miss the cycle.

Reviewed By: mroch

Differential Revision: D3345572

fbshipit-source-id: dd21d6bdcac162f577542a2eaea7253fc672d20e