don't add properties to dictionary types
Summary:
Setting a property on an object type with a dictionary would add the property to
the object type if it didn't already exist. The motivation was to track the type
of the property more precisely.

Unfortunately, this led to some surprising consequences: we would report some
errors only when we saw constraints in a particular order. Since this order is
in general not correlated to the order in which code gets executed at run time,
we consider "races" like this to be bugs.

Worse, considering dictionary types to be unsealed like this would sometimes
cause us to miss real type incompatibility errors: a set along one path would
silence errors on gets along different paths.

This diff fixes by closely following what happens with dictionaries on gets.

Fixes https://github.com/facebook/flow/issues/1745

Reviewed By: samwgoldman

Differential Revision: D3322062

fbshipit-source-id: e37ac3eb8459b4a176575d2f4dfa35a9eb02ad53