[hack] Fix hh_server execution on Mac
Summary:
This is a partial revert of D3264178. AFAICT this change is unrelated to the issue at hand there (hhi files in bytecode mode) but was snuck in for something to do with reproducible builds. I'm still not entirely sure what this was doing, but the options in question don't work on OS X's bsdtar, and in fact cause the typechecker to crash at startup there. (I reverted another reproducbile build change for the same reason in D3219053.)

If someone wants to work on reproducible builds and needs info/testing on a Mac system, LMK. Happy to help review. But until then I value a //working// build everywhere more than a reproducible one in some places.

Reviewed By: dlreeves

Differential Revision: D3336338

fbshipit-source-id: d7d65f6f4e2183bccb462bf2cb1ef08a1c8bb517