Missing .exp file error should indicate the file's intended name
Summary:
This has been driving me nuts for the last few days (not because of the goof itself, but because I authored the goof).  When a test directory's .exp file is missing, the error message looks like "Missing tests/some_test/.exp or .flowconfig...."  This is slightly unclear to a new dev and it looks like hell.
Closes https://github.com/facebook/flow/pull/1837

Differential Revision: D3352271

Pulled By: mroch

fbshipit-source-id: 04ec7e402b52848cd9fe36fa757d885aecfd8250